v2.1.0 - Added anti-flicker, code works correctly for the subpages, added files for the Color Meaning page
Anti-flicker - when the page loads, the background image took some time to load (<0.1s), so the gradient was visible for a really short time. Now, the background has a color added on page load, so the bg image doesn't flicker.

Font path was changed according to the current page, so now it loads correctly.
Additionally, the ucp-(color model) class could be added to a non-existent element, now it's fixed.

Added HTML and CSS files for the Color Meaning page.