v2.4.5 - Fixed a bug with PDF height
The height for the layout parent element had dynamically assigned height based on the currently displayed layout.
When printing PDF, it would set a height to that element, but it would be a wrong one since the screen size is different (user screen vs PDF size).

Now the height is not set in TS, but it's set to auto in css. HAd to change some other things, but now the height is set correctly everywhere.

Additionally, I started working on the second layout.