v2.6.0 - Navigation fixes + Color Tips page
Added the Color Tips page (no real content at the moment - just some basic divs).

Fixed and improved the page navigation.

Improved nav animation - now, the page title won't have the jump animation when the user hovers over the current page "dot" element.

Fixed the issue with the last page not redirecting to the first page when clicking right arrow and not showing the next (so the first) page in the dot elements.