v2.7.0 - Typefaces page, design changes, cohesiveness improvement
Added Typefaces page. It introduces the main and secondary typeface to the user. The page contains short description and use proposition.

Changed the typefaces across the whole project (mainly the Layout page) to the "client's" typefaces.
Added a class that can be added. If so, it makes the font size smaller in the typeface doesn't have a normal size.

Also, some of the page elements were not standardized across the pages, but now they are.