v.2.9.2 - Typeface paths and ReadMe screenshot
Changed the typeface paths, because the old ones ("Fonts" instead of "fonts") were causing some problems when hosting the website on github.

Additionally, I changed the screenshot of the Color Chart page. There was an issue when converting PDF to PNG and a few words were not displayed (there was a gradient instead).