v2.9.4 - PDF Print, Dark Mode Fixes
Changed the "background color" tile on the main page to remove the flashing white color (dark mode).

Fixed an issue with PDFs not rendering correctly. The background image and gradient were not covering the entire page.
There might have been a change in the print() functionality that caused this issue.

Fixed the Contact Page colors in the dark mode.

Implemented other minor design changes.